+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Remove "save screen" option (always on)
 * Change again the threshold for triggering a non-draw frame
 * Don't scroll input editor when rewinding (#381)
+* Raise the limit of thread number in states
 
 ### Fixed